+++ client.go
@@ -52,8 +52,9 @@ type Client struct {
 	VatCodeInfoList      *VatCodeInfoListService
 	ProductGroupNVL      *ProductGroupNVLService
 	ProductGroup         *ProductGroupService
-	OrganizationInfoList *OrganizationInfoListService
 	Organization         *OrganizationService
+	OrganizationInfo     *OrganizationInfoService
+	OrganizationInfoList *OrganizationInfoListService
 }
 
 // RequestCompletionCallback defines the type of the request callback function
@@ -87,8 +88,9 @@ func NewClient(httpClient *http.Client, baseURL *url.URL) *Client {
 	c.VatCodeInfoList = NewVatCodeInfoListService(c)
 	c.ProductGroupNVL = NewProductGroupNVLService(c)
 	c.ProductGroup = NewProductGroupService(c)
-	c.OrganizationInfoList = NewOrganizationInfoListService(c)
 	c.Organization = NewOrganizationService(c)
+	c.OrganizationInfo = NewOrganizationInfoService(c)
+	c.OrganizationInfoList = NewOrganizationInfoListService(c)
 
 	// Commands
